Tweak the InitialTestRunProcessFactory
